+++ Plugins/ActorPools/Source/NexusActorPools/Public/NActorPoolsSettings.h
@@ -4,6 +4,7 @@
 #pragma once
 
 #include "InputCoreTypes.h"
+#include "NActorPoolSet.h"
 #include "NActorPoolSettings.h"
 #include "NSettingsUtils.h"
 #include "Macros/NSettingsMacros.h"
@@ -41,4 +42,12 @@ public:
 	UPROPERTY(Config, EditAnywhere, BlueprintReadOnly, Category = "Actor Pools", DisplayName ="Default Settings",
 		meta=(ToolTip="The default settings applied to a created NActorPool when no settings are provided."))
 	FNActorPoolSettings DefaultSettings;
+	
+	UPROPERTY(Config, EditAnywhere, BlueprintReadOnly, Category="Initialization", DisplayName="Always Create",
+		meta=(ToolTip="Always create the outlined actor pool sets when a world is created."))
+	TArray<TSoftObjectPtr<UNActorPoolSet>> AlwaysCreateSets;
+	
+	UPROPERTY(Config, EditAnywhere, BlueprintReadOnly, Category = "Initialization", DisplayName = "Ignore World Prefixes", 
+		meta=(Tooltop="Ignore attempting to auto create pools for worlds thats name starts with or is outlined."))
+	TArray<FString> IgnoreWorldPrefixes;
 };